# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WiseEff API",
    "version": "m5"
  },
  "paths": {
...
    "/api/v1/operations/pilot-readiness": {
      "get": {
        "operationId": "operations.pilotReadiness",
        "summary": "Pilot readiness gate",
        "tags": [
          "operations"
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilotReadinessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "PilotReadinessResponse": {
        "type": "object",
        "x-wiseeff-schema": "PilotReadinessResponse"
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "WiseEff error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}